# openapi.yaml
openapi: 3.0.1
info:
  title: FilmArchivePlanner (FAP)
  description: "REST API for the FAP backend."
  version: 0.1.0
...
servers:
  - url: http://localhost:3000
...
paths:
...
  /tags:
    get:
      tags:
        - tags
      summary: Get all tags
      description: Get all tags from the archive. Query parameters can be used to return only certain tags.
      operationId: getTags
      parameters:
        - in: query
          name: searchTagPagingDto
          schema:
            $ref: "#/components/schemas/SearchTagPagingDto"
          description: All properties of searchTagPagingDto are valid parameters see specifaction for details. The searchString parameter disables all detailed field search parameters.
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TagDto"
        400:
          description: Invalid input
          content: {}
        401:
          description: Unauthorized
          content: {}
        403:
          description: Missing permission
          content: {}
    post:
      tags:
        - tags
      summary: Add new tag
      description: Adds a new tag to the archive
      operationId: addTag
      requestBody:
        description: Tag object that needs to be added to the store
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUpdateTagDto"
        required: true
      responses:
        201:
          description: Successful creation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagDto"
        400:
          description: Invalid input
          content: {}
        401:
          description: Unauthorized
          content: {}
        403:
          description: Missing permission
          content: {}
      x-codegen-request-body-name: body
  /tags/{id}:
    get:
      tags:
        - tags
      summary: Get tag
      description: Returns a single tag with the specified ID
      operationId: getTagById
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagDto"
        400:
          description: Invalid ID supplied
          content: {}
        401:
          description: Unauthorized
          content: {}
        403:
          description: Missing permission
          content: {}
        404:
          description: Not found
          content: {}
    put:
      tags:
        - tags
      summary: Update tag
      description: Updates the tag with the specified ID
      operationId: updateTag
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUpdateTagDto"
      responses:
        204:
          description: Operation successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TagDto"
        400:
          description: Invalid input
          content: {}
        401:
          description: Unauthorized
          content: {}
        403:
          description: Missing permission
          content: {}
        404:
          description: Not found
          content: {}
    delete:
      tags:
        - tags
      summary: Delete tag
      description: Deletes the tag with the specified ID
      operationId: deleteTag
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        400:
          description: Invalid ID supplied
          content: {}
        401:
          description: Unauthorized
          content: {}
        403:
          description: Missing permission
          content: {}
        404:
          description: Not found
          content: {}
  /tags/type/{type}:
    get:
      tags:
        - tags
      summary: Get tags by type
      description: Returns a list of tag with the specified tag type
      operationId: getTagByType
      parameters:
        - $ref: "#/components/parameters/Type"
      responses:
        200:
          description: successful operation
          content:
            application/json:
              "schema":
                {
                  "type": "array",
                  "items": { $ref: "#/components/schemas/TagDto" },
                }
        400:
          description: Invalid type supplied
          content: {}
        401:
          description: Unauthorized
          content: {}
        403:
          description: Missing permission
          content: {}
        404:
          description: Not found
          content: {}
...
components:
  parameters:
    Id:
      name: id
      in: path
      description: ID of entity to return
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
...
    Type:
      name: type
      in: path
      description: TagType #TODO: change TagType to Enum
      required: true
      schema:
        format: string
  schemas:
    DatabaseEntrySchema:
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
        created_at:
          type: string
        last_updated:
          type: string
    PagingSchema:
      properties:
        page:
          description: Used for pagination. The number of the page to return.
          type: integer
          minimum: 1
        limit:
          description: Used for pagination. The number of items to return.
          type: integer
          minimum: 1
        sortOrder:
          description: Used for pagination. The number of items to return.
          type: string
        orderBy:
          description: Used for pagination. The number of items to return.
          type: string
        searchString:
          description: Searches all string fields containing the searchString and returns every match. If specified, detailed search fields are ignored.
          type: string
...
    TagSchema:
      properties:
        type:
          type: string
        value:
          type: string
        user:
          type: string # TODO change to user type
        public:
          type: boolean
...
    CreateUpdateTagDto:
      type: object
      required:
        - type
        - name
      allOf:
        - $ref: "#/components/schemas/TagSchema"
    TagDto:
      type: object
      allOf:
        - $ref: "#/components/schemas/DatabaseEntrySchema"
        - $ref: "#/components/schemas/CreateUpdateTagDto"
    SearchTagPagingDto:
      type: object
      allOf:
        - $ref: "#/components/schemas/PagingSchema"
        - $ref: "#/components/schemas/TagSchema"
...